backend/requirements.txt
django
djangorestframework
djoser

django-polymorphic
pillow
drf-extra-fields
django-filter
drf-spectacular
django-treebeard
python-dotenv
django-phonenumber-field[phonenumbers]
pytils
pytz
timezonefinder

psycopg2-binary
django-dbbackup

django-storages
boto3

gunicorn
django-cors-headers

celery[redis]
django-celery-beat
redis
django-redis

django_debug_toolbar